Factored out permission_state_invocation_results API to GlobalScope
Instead of Window
