+++ components/script/dom/permissions.rs
@@ -145,7 +145,6 @@ impl Permissions {
                         // (Revoke) Step 3.
                         let globalscope = self.global();
                         globalscope
-                            .as_window()
                             .permission_state_invocation_results()
                             .borrow_mut()
                             .remove(&root_desc.name.to_string());
@@ -178,7 +177,6 @@ impl Permissions {
                         // (Revoke) Step 3.
                         let globalscope = self.global();
                         globalscope
-                            .as_window()
                             .permission_state_invocation_results()
                             .borrow_mut()
                             .remove(&root_desc.name.to_string());
@@ -272,7 +270,6 @@ impl PermissionAlgorithm for Permissions {
                 let prompt = format!("{} {} ?", REQUEST_DIALOG_MESSAGE, perm_name.clone());
                 let state = prompt_user_from_embedder(prompt, &globalscope);
                 globalscope
-                    .as_window()
                     .permission_state_invocation_results()
                     .borrow_mut()
                     .insert(perm_name.to_string(), state);
@@ -312,7 +309,6 @@ pub fn get_descriptor_permission_state(
             PermissionState::Granted
         } else {
             globalscope
-                .as_window()
                 .permission_state_invocation_results()
                 .borrow_mut()
                 .remove(&permission_name.to_string());
@@ -324,7 +320,6 @@ pub fn get_descriptor_permission_state(
 
     // Step 3.
     if let Some(prev_result) = globalscope
-        .as_window()
         .permission_state_invocation_results()
         .borrow()
         .get(&permission_name.to_string())
@@ -334,7 +329,6 @@ pub fn get_descriptor_permission_state(
 
     // Store the invocation result
     globalscope
-        .as_window()
         .permission_state_invocation_results()
         .borrow_mut()
         .insert(permission_name.to_string(), state);
